# Add AVX2 Support