- Add `ADA_AVX2` macro in `include/ada/common_defs.h` enabled if `defined(__AVX2__)`
- Implement `has_tabs_or_newline` using AVX2